Added Groups Model, Controller, and View
Added the Model for Groups and included a function to generate random and unique tokens.  Added the Controller for Groups with functions new and create.  Added a View for Groups called new.html.erb where a logged in user is able to create a group and is then added to the group created.  Also, added current_group to application controller.  Added some comments to the code I created.